Refactoring matchCreateForm.jsx and profileEditForm.jsx
